Backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv
requests
docker
langchain
langchain-community
langchain-google-genai>=1.0.3
langgraph